Changed my code for Level 4
Hey Nicole! Your level 1 is gone with the changes you made last night?
I made my code work with the new classes you made for the players
(thank you!!), but couldn’t get it to run with the minim and without
Level 1 so I commented that out until we can make changes to it to get
it to work. Is Level 1 supposed to be missing or is that a problem?